feat: Move Tutors page, triple-fusions, responsive design, IF-style visual overhaul
Backend:
- New endpoints GET /moves/tutors/all and /moves/experts/all (frontend bulk fetch)
- /moves/{id}/tutors now returns move_name_en + move_name_fr
- search_pokemon_locations tool (gift/trade/static/wild lookup in pokemon_location)
- wiki_service: fetch full page when intro < 300 chars (was < 1)
- triple_fusion_service: use type IDs directly (was decomposing compound names → wrong ×8 values)
- system.md: anti-loop rules, search_pokemon_locations guidance, key game mechanics section

Frontend:
- New page /moves/tutors — 41 classic tutors grouped by location + Move Experts by island
- New page /triple-fusions — 23 triple fusions with types and weaknesses
- FusionSelector: GameFilter (Kanto / Hoenn / All) above pickers
- Navbar: hamburger + full-width drawer on mobile (md:hidden)
- Full responsive pass: tables hidden sm:table-cell, panels flex-col md:flex-row, SpriteCard w-32 sm:w-40
- IF-style design: navy #090c1a bg, gold #e8b84b accents, @theme CSS tokens, grid texture
- TypeBadge: box-shadow glow + text-shadow
- PokemonCard: type gradient, top accent line, sprite radial glow, dynamic hover border
- StatBar: optional typeColor prop, gradient bar with glow
- Pokedex detail header: card with type gradient + accent line, gold tabs
- EvolutionChain: smaller nodes (w-14), word-wrap condition, gold active border

+++ backend/schemas/weakness.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from pydantic import BaseModel
 
 
@@ -8,6 +6,6 @@ class WeaknessOut(BaseModel):
     attacking_type_id: int
     attacking_type_name_en: str
     attacking_type_name_fr: str | None
-    multiplier: Decimal
+    multiplier: float
 
     model_config = {"from_attributes": True}
